tensorflow inference and training tests refactored
